Separate TPA adjustment for YAW axis.
Allows configuring a separate TPA for YAW axis. This is useful especially in tricopters.

The CLI parameters to configure this are:
tpa_yaw_breakpoint - similar to regular TPA
tpa_yaw_rate - now this is bit different as it supports reducing and increasing the PID values. Normal TPA only allows reduction. Value 100 (%) here means that no TPA is applied. Value 50 means that at maximum throttle the yaw PID gains are reduced to half. Value 200 % means that at maximum throttle the gains are doubled. Throttle values in between the range are interpolated.